Improved tests to require explicit db config and share connection fixtures.
